+++ cameraview/build.gradle
@@ -32,7 +32,6 @@ android {
     }
 
     sourceSets {
-        main.java.srcDirs += 'src/main/views'
         main.java.srcDirs += 'src/main/utils'
         main.java.srcDirs += 'src/main/gles'
     }
