+++ cameraview/src/main/java/com/otaliastudios/cameraview/GridLinesLayout.java
