+++ cameraview/src/main/java/com/otaliastudios/cameraview/RendererThread.java
